feat: create folder catalogue and move files DownloadableCVCard and Modal
